mtree: fix warning/error produced by SunC

+++ mtree/crc.c
@@ -114,7 +114,7 @@ static const uint32_t crctab[] = {
  * locations to store the crc and the number of bytes read.  It returns 0 on
  * success and 1 on failure.  Errno is set on failure.
  */
-uint32_t crc_total = ~0;		/* The crc over a number of files. */
+uint32_t crc_total = ~(uint32_t)0;		/* The crc over a number of files. */
 
 int
 crc(int fd, uint32_t *cval, uint32_t *clen)
